requirements.txt
anyascii==0.3.2
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==5.0.9
django-classy-tags==4.1.0
django-cms==4.1.3
django-cors-headers==4.5.0
django-filter==24.3
django-formtools==2.5.1
django-permissionedforms==0.1
django-sekizai==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-text-ckeditor==5.1.6
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
l18n==2021.3
laces==0.1.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pillow_heif==0.18.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
telepath==0.3.1
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Willow==1.8.0
